Refactors route and controller
