--- src/scripts/pages/yourCases.ts
@@ -1,14 +1,0 @@
-/**
- * Your Cases Page Configuration
- *
- * Handles client-side functionality for the cases page
- */
-
-import { devLog } from '../helpers/index.js';
-
-document.addEventListener('DOMContentLoaded', function () {
-  devLog('Your Cases: Page loaded');
-
-  // No tab management needed - server-side rendering handles tab visibility
-  // Add any other client-side functionality here as needed
-});
